Preloader and Button Hover interaction complete
I was playing around with the code for a while, trying to figure out why the hover interaction wasn't working and found that the JQuery link was placed after the call to the js folder...apparently the link needs to go before. Nonetheless, that problem was fixed and the hover now works. Future iterations should include centering the text on the button. I also managed to fix syntax errors that allow for the preloader to bounce around for about 3000ms and move onto the main content.